proyectoDjangu/app/serializers.py
from .models import Producto, Marca
from rest_framework import serializers

class MarcaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Marca
        field = '__all__'

class ProductoSerializer(serializers.ModelSerializer):
    nombre_marca = serializers.CharField(read_only=True, source="marca")
    class Meta:
        model = Producto
        fields = '__all__'
